Refactor and clean up headers, naming, and file structure
Removed unused main.h and its references, updated function and enum naming to follow consistent conventions (e.g., PascalCase for methods, uppercase for enum values), and cleaned up includes in several headers. Adjusted method signatures and usage in soundTrack, renderer, and textLoader for clarity and consistency. Minor code cleanups and comments added for demo-specific code.

+++ include/syncTracker.h
@@ -19,9 +19,9 @@ public:
 	explicit SyncTracker();
 	~SyncTracker() noexcept;
 
-	void Update(SoundTrack& SoundTrack);
-	float FetchValue(const std::string& name);
-	float CurrentTime();
+	void	Update(SoundTrack& SoundTrack);
+	double	FetchValue(const std::string& name);
+	float	CurrentTime();
 
 };
